+++ src/AnimeParams.ts
@@ -8,6 +8,9 @@ export class CellValue {
 	cellName: string;
 }
 
+/**
+ * カーブの補間パラメータ。
+ */
 export class IpCurve {
 	values: number[] = [];
 }
@@ -33,6 +36,11 @@ export const ipTypes = [
  */
 export type IpType = typeof ipTypes[number];
 
+/**
+ * キーフレーム。
+ *
+ * 描画に関する属性のある時刻の値と補間方法を表す。
+ */
 export class KeyFrame<T> {
 	time: number;
 	value: T;
@@ -40,16 +48,31 @@ export class KeyFrame<T> {
 	ipCurve: IpCurve;
 }
 
+/**
+ * カーブ。
+ *
+ * 描画に関する属性の時間に伴う変化を表す。
+ */
 export class Curve<T> {
 	attribute: string = "";
 	keyFrames: KeyFrame<T>[] = [];
 }
 
+/**
+ * カーブタイ。
+ *
+ * 複数のカーブをまとめ、ボーンに対応づける。
+ */
 export class CurveTie {
 	boneName: string = "";
 	curves: Curve<any>[] = [];
 }
 
+/**
+ * アニメーション。
+ *
+ * 各ボーンのアニメーションを保持する。
+ */
 export class Animation {
 	name: string = "";
 	fps: number = 0;