+++ src/serialize/ProjectV2.ts
@@ -1,16 +1,40 @@
 import type { Schema } from "./Schema";
 
+/**
+ * プロジェクト(V2形式）。
+ *
+ * プロジェクトはさまざまなアニメーションデータのメタデータを持つ。
+ */
 export interface ProjectV2 {
+	/**
+	 * boneSetファイル名のリスト。
+	 */
 	boneSetFileNames: string[];
+
+	/**
+	 * skinファイル名のリスト。
+	 */
 	skinFileNames: string[];
+
+	/**
+	 * animationファイル名のリスト。
+	 */
 	animationFileNames: string[];
+
+	/**
+	 * effectファイル名のリスト。
+	 */
 	effectFileNames: string[];
+
+	/**
+	 * ユーザーデータ。
+	 */
 	userData?: any;
 
 	/**
-	 * スキーマ
+	 * スキーマ。
 	 *
-	 * プロジェクト配下のファイルのスキーマ。
+	 * 各種データの構造を定義するデータ。
 	 */
 	schema?: Schema;
 }