+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## unreleased changes
+
+* `Resource#loadProject()` に `g.AssetAccessor` (`scene.asset`, `game.asset`) を渡せるように修正
+* Akashic Animation のデータの容量を削減する ArrayOrientedPorter を追加
+
 ## 4.1.0
 
 * SpriteStudioで利用される補間方法の一部をサポート